Higher-order choreographies work!
Major change to how projection works: when you project a =let=
construct for the label that does *not* have the variable, instead of
projecting to a fresh global function, we project to =with= like case
of the label with the variable, and simply ignore the result of rhs of
the assignment.

This is important because it means variables stay in scope through the
body of the function. A problem that cropped up was that the book
seller example would have `the_book` out-of-scope in the body of the
`with` because, the way this was projected originally, that would all
be in a separate function.

This is a problem for Elixir and not for Pirouette because Elixir does
not support nested functions, except as lambda expressions. But those
are not recursive, so boo.