Checkpoint: draft of splatting/unsplatting variables in functions
There will likely need to be some serious modifications to this. In
particular, I am not sure if I will need to thread variables around.
For example, I run projection on the body of the function—might that
reveal extra variables or split into a *new* function?!

It might be OK though—with blocks are self-contained (right?) and will
push the variable they bind deeper into the context. But when I do a
receive all I need to unsplat are the variables *seen until this
point*—the new variables in the pattern will appear in the *next*
function.